refactor(operator_commands_dashboard): extract metrics/costs/budget to monitoring module
Move three small read-only HTTP handlers (metrics, costs, get_budget,
set_budget) plus the budget_config_path helper into a new sibling
monitoring.rs module.

- routes.rs: 5558 -> 5488
- monitoring.rs (NEW, 74)

Refs #1266
